+++ src/os2datascanner/engine2/conversions/__init__.py
@@ -9,6 +9,7 @@ from .registry import convert, conversion_exists  # noqa
 from . import text  # noqa
 from . import fallback  # noqa
 from . import last_modified  # noqa
+from . import size  # noqa
 from . import image_dimensions  # noqa
 from . import get_links  # noqa
 from . import manifest  # noqa